pyproject.toml
[project]
name = "zey-batch"
version = "0.1.0"
description = "SMS Campaign Management System for Customer Engagement"
requires-python = ">=3.11"
dependencies = [
    "twilio>=8.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "colorama>=0.4.6",
    "rich>=13.0.0",
    "python-dateutil>=2.8.0",
]

[project.scripts]
sms-campaign = "sms_campaign.cli:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sms_campaign"]

[tool.hatch.build]
exclude = [
  "data/*.xlsx",
  "data/*.csv",
  "data/archive",
  "data/delete_temp",
  "logs",
]
